Added Cpb file generation engine tests. Moved from schema to code (#613)
